Changes from review

+++ vagrant/create_scion_box.sh
@@ -6,16 +6,22 @@ VERSION="0.0.1.2"
 BASE=$(realpath $(dirname "$0"))
 cd "$BASE"
 
+do_create=1
 do_package=0
 usage="$(basename "$0") [-p]
 
 where:
     -h          this help
-    -p          also package the VM"
+    -p          also package the VM
+    -j          only package the VM (don't create it)"
 
-while getopts ":hp" opt; do
+while getopts ":hpj" opt; do
   case $opt in
     p)
+      do_package=1
+      ;;
+    j)
+      do_create=0
       do_package=1
       ;;
     h)
@@ -38,15 +44,18 @@ done
 # running this script will create a new vagrant virtual machine with scion installed and ready to work
 # if do_package==1, it will also run `vagrant package` to get a .box file that can be uploaded to vagrant cloud
 # for more info see https://www.vagrantup.com/docs/vagrant-cloud/boxes/create-version.html
-cp ../scion_install_script.sh .
-vagrant destroy -f
-VAGRANT_VAGRANTFILE=Vagrantfile-bootstrap vagrant destroy -f
-echo '------------------------------------ updating vagrant boxes'
-vagrant box update
-echo '------------------------------------ creating bootstrap vagrant VM'
-VAGRANT_VAGRANTFILE=Vagrantfile-bootstrap vagrant up
-echo '------------------------------------ creating base vagrant VM'
-vagrant up
+if [ $do_create -eq 1 ]; then
+  echo '------------------------------------ destroying old VMs'
+  cp ../scion_install_script.sh .
+  vagrant destroy -f
+  VAGRANT_VAGRANTFILE=Vagrantfile-bootstrap vagrant destroy -f
+  echo '------------------------------------ updating vagrant boxes'
+  VAGRANT_VAGRANTFILE=Vagrantfile-bootstrap vagrant box update
+  echo '------------------------------------ creating bootstrap vagrant VM'
+  VAGRANT_VAGRANTFILE=Vagrantfile-bootstrap vagrant up
+  echo '------------------------------------ creating base vagrant VM'
+  vagrant up --provision
+fi
 
 if [ $do_package -eq 1 ]; then
     echo '------------------------------------ packaging base vagrant VM'
@@ -54,11 +63,10 @@ if [ $do_package -eq 1 ]; then
     vagrant package --output scion-base.box --vagrantfile Vagrantfile-exported
     vagrant destroy -f
     echo '------------------------------------ locally adding base vagrant VM'
-    vagrant box remove -f scion/ubuntu-16.04-64-scion --box-version "$VERSION" || true
-    cp metadata-template.json metadata.json
-    sed -i -- "s|_VERSION_|$VERSION|g" metadata.json
-    sed -i -- "s|_DIRFULLPATH_|$BASE|g" metadata.json
-    vagrant box add metadata.json
-    echo '------------------------------------'
+    vagrant box remove -f scion/ubuntu-16.04-64-scion --all || true
+    cp metadata-template.json /tmp/metadata.json
+    sed -i -- "s|_VERSION_|$VERSION|g" /tmp/metadata.json
+    sed -i -- "s|_DIRFULLPATH_|$BASE|g" /tmp/metadata.json
+    vagrant box add /tmp/metadata.json
 fi
 echo "Done."
